// apps/api/src/shared/wordSync.ts
export interface LyricWord {
  readonly text: string;
  readonly start: number;
  readonly end: number;
}